pyproject.toml
[tool.poetry]
name = "live-coding"
version = "0.1.0"
description = ""
authors = ["vilelf <vilelf.dev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.61.2"
requests = "^2.24.0"
alembic = "^1.4.3"
FastAPI-SQLAlchemy = "^0.2.1"
python-dotenv = "^0.15.0"
psycopg2-binary = "^2.8.6"
fastapi-utils = "^0.2.1"

[tool.poetry.dev-dependencies]
uvicorn = "^0.12.2"
prospector = "^1.3.1"
pytest = "^6.1.2"
autopep8 = "^1.5.4"
pytest-cov = "^2.10.1"
mypy = "^0.790"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
